+++ nbtc/relayer.go
@@ -1,28 +1,29 @@
 package nbtc
 
 import (
-	"bytes"
+	"context"
 	"errors"
-	"fmt"
 	"time"
 
-	"github.com/btcsuite/btcd/chaincfg/chainhash"
-	"github.com/btcsuite/btcd/rpcclient"
-	"github.com/btcsuite/btcd/wire"
-	"github.com/gonative-cc/relayer/bitcoin"
 	"github.com/gonative-cc/relayer/dal"
+	err "github.com/gonative-cc/relayer/errors"
+	"github.com/gonative-cc/relayer/ika2btc"
+	"github.com/gonative-cc/relayer/native2ika"
 	"github.com/mattn/go-sqlite3"
 	"github.com/rs/zerolog/log"
 )
 
-// Relayer broadcasts pending transactions from the database to the Bitcoin network.
+// Relayer handles the flow of transactions from the Native chain to Bitcoin.
+// It uses two processors:
+// - nativeProcessor: To send transactions from Native to IKA for signing.
+// - btcProcessor: To broadcast signed transactions to Bitcoin and monitor confirmations.
 type Relayer struct {
-	db                      *dal.DB
-	btcClient               bitcoin.Client
-	shutdownChan            chan struct{}
-	processTxsTicker        *time.Ticker
-	confirmTxsTicker        *time.Ticker
-	txConfirmationThreshold uint8
+	db               *dal.DB
+	nativeProcessor  *native2ika.Processor
+	btcProcessor     *ika2btc.Processor
+	shutdownChan     chan struct{}
+	processTxsTicker *time.Ticker
+	confirmTxsTicker *time.Ticker
 }
 
 // RelayerConfig holds the configuration parameters for the Relayer.
@@ -32,27 +33,29 @@ type RelayerConfig struct {
 	ConfirmationThreshold uint8         `json:"confirmationThreshold"`
 }
 
-// NewRelayer creates a new Relayer instance with the given configuration.
+// NewRelayer creates a new Relayer instance with the given configuration and processors.
 func NewRelayer(
-	btcClientConfig rpcclient.ConnConfig,
 	relayerConfig RelayerConfig,
 	db *dal.DB,
+	nativeProcessor *native2ika.Processor,
+	btcProcessor *ika2btc.Processor,
 ) (*Relayer, error) {
 
 	if db == nil {
-		err := fmt.Errorf("database cannot be nil")
+		err := err.ErrNoDB
 		log.Err(err).Msg("")
 		return nil, err
 	}
 
-	if btcClientConfig.Host == "" || btcClientConfig.User == "" || btcClientConfig.Pass == "" {
-		err := fmt.Errorf("missing bitcoin node configuration")
+	if nativeProcessor == nil {
+		err := err.ErrNoNativeProcessor
 		log.Err(err).Msg("")
 		return nil, err
 	}
 
-	client, err := rpcclient.New(&btcClientConfig, nil)
-	if err != nil {
+	if btcProcessor == nil {
+		err := err.ErrNoBtcProcessor
+		log.Err(err).Msg("")
 		return nil, err
 	}
 
@@ -69,37 +72,29 @@ func NewRelayer(
 	}
 
 	return &Relayer{
-		db:                      db,
-		btcClient:               client,
-		shutdownChan:            make(chan struct{}),
-		processTxsTicker:        time.NewTicker(relayerConfig.ProcessTxsInterval),
-		confirmTxsTicker:        time.NewTicker(relayerConfig.ConfirmTxsInterval),
-		txConfirmationThreshold: relayerConfig.ConfirmationThreshold,
+		db:               db,
+		nativeProcessor:  nativeProcessor,
+		btcProcessor:     btcProcessor,
+		shutdownChan:     make(chan struct{}),
+		processTxsTicker: time.NewTicker(relayerConfig.ProcessTxsInterval),
+		confirmTxsTicker: time.NewTicker(relayerConfig.ConfirmTxsInterval),
 	}, nil
 }
 
-// Start starts the relayer's main loop to broadcast transactions.
-func (r *Relayer) Start() error {
+// Start starts the relayer's main loop.
+func (r *Relayer) Start(ctx context.Context) error {
 	for {
 		select {
 		case <-r.shutdownChan:
-			r.btcClient.Shutdown()
+			r.btcProcessor.Shutdown()
 			return nil
 		case <-r.processTxsTicker.C:
-			if err := r.processPendingTxs(); err != nil {
-				var sqliteErr *sqlite3.Error
-				//TODO: decide on which exact errors to continue and on which to stop the relayer
-				if errors.As(err, &sqliteErr) {
-					log.Err(err).Msg("Critical error updating transaction status, shutting down")
-					close(r.shutdownChan)
-				} else {
-					log.Err(err).Msg("Error processing transactions, continuing")
-				}
-			}
+			_ = r.processSignRequests(ctx)
+			_ = r.processSignedTxs()
 		//TODO: do we need a subroutine for it? Also i think there  might be a race condition on the database
 		// so probably we should wrap the db in a mutex
 		case <-r.confirmTxsTicker.C:
-			if err := r.checkConfirmations(); err != nil {
+			if err := r.btcProcessor.CheckConfirmations(); err != nil {
 				var sqliteErr *sqlite3.Error
 				if errors.As(err, &sqliteErr) {
 					log.Err(err).Msg("Critical error updating transaction status, shutting down")
@@ -112,60 +107,34 @@ func (r *Relayer) Start() error {
 	}
 }
 
-// processPendingTxs processes pending transactions from the database.
-func (r *Relayer) processPendingTxs() error {
-	pendingTxs, err := r.db.GetPendingTxs()
-	if err != nil {
+// processSignRequests processes signd requests from the Native chain.
+func (r *Relayer) processSignRequests(ctx context.Context) error {
+	if err := r.nativeProcessor.Run(ctx); err != nil {
+		var sqliteErr *sqlite3.Error
+		//TODO: decide on which exact errors to continue and on which to stop the relayer
+		if errors.As(err, &sqliteErr) {
+			log.Err(err).Msg("Critical error with the database, shutting down")
+			close(r.shutdownChan)
+		} else {
+			log.Err(err).Msg("Error processing native transactions, continuing")
+		}
 		return err
-	}
-
-	for _, tx := range pendingTxs {
-		var msgTx wire.MsgTx
-		if err := msgTx.Deserialize(bytes.NewReader(tx.RawTx)); err != nil {
-			return err
-		}
-
-		txHash, err := r.btcClient.SendRawTransaction(&msgTx, false)
-		if err != nil {
-			return fmt.Errorf("error broadcasting transaction: %w", err)
-		}
-
-		err = r.db.UpdateTxStatus(tx.BtcTxID, dal.StatusBroadcasted)
-		if err != nil {
-			return fmt.Errorf("DB: can't update tx status: %w", err)
-		}
-
-		log.Info().Str("txHash", txHash.String()).Msg("Broadcasted transaction: ")
 	}
 	return nil
 }
 
-// checkConfirmations checks all the broadcasted transactions to bitcoin
-// and if confirmed updates the database accordingly.
-func (r *Relayer) checkConfirmations() error {
-	broadcastedTxs, err := r.db.GetBroadcastedTxs()
-	if err != nil {
+// processSignedTxs processes signed transactions and broadcasts them to Bitcoin.
+func (r *Relayer) processSignedTxs() error {
+	if err := r.btcProcessor.Run(); err != nil {
+		var sqliteErr *sqlite3.Error
+		//TODO: decide on which exact errors to continue and on which to stop the relayer
+		if errors.As(err, &sqliteErr) {
+			log.Err(err).Msg("Critical error updating transaction status, shutting down")
+			close(r.shutdownChan)
+		} else {
+			log.Err(err).Msg("Error processing bitcoin transactions, continuing")
+		}
 		return err
-	}
-
-	for _, tx := range broadcastedTxs {
-		hash, err := chainhash.NewHash(tx.Hash)
-		if err != nil {
-			return err
-		}
-		txDetails, err := r.btcClient.GetTransaction(hash)
-		if err != nil {
-			return fmt.Errorf("error getting transaction details: %w", err)
-		}
-
-		// TODO: decide what threshold to use. Read that 6 is used on most of the cex'es etc.
-		if txDetails.Confirmations >= int64(r.txConfirmationThreshold) {
-			err = r.db.UpdateTxStatus(tx.BtcTxID, dal.StatusConfirmed)
-			if err != nil {
-				return fmt.Errorf("DB: can't update tx status: %w", err)
-			}
-			log.Info().Msgf("Transaction confirmed: %s", tx.Hash)
-		}
 	}
 	return nil
 }